+++ src/modules/CC1101/CC1101.h
@@ -950,7 +950,7 @@ class CC1101: public PhysicalLayer {
     int16_t config();
     int16_t directMode();
     static void getExpMant(float target, uint16_t mantOffset, uint8_t divExp, uint8_t expMax, uint8_t& exp, uint8_t& mant);
-    int16_t setPacketMode(uint8_t mode, uint8_t len);
+    int16_t setPacketMode(uint8_t mode, uint16_t len);
 };
 
 #endif
